Shorten long blocks to new standard of 80 lines
