+++ app/interfaces/api/v2/case_workers/claim.rb
@@ -7,22 +7,21 @@ module API
         namespace :case_workers do
           params do
             optional :api_key, type: String, desc: 'REQUIRED: The API authentication key of the user'
-            optional :status,
-                     type: String,
-                     values: %w[current allocated unallocated archived],
-                     default: 'current',
-                     desc: 'REQUIRED: Only returns claims in the specified status'
-            optional :scheme,
-                     type: String,
-                     values: %w[agfs lgfs],
-                     default: 'agfs',
-                     desc: 'OPTIONAL: This will be used to filter the list of allocated/unallocated claims'
-            optional :filter,
-                     type: String,
-                     values: %w[all redetermination awaiting_written_reasons fixed_fee cracked trial guilty_plea
-                                graduated_fees interim_fees warrants interim_disbursements risk_based_bills],
-                     default: 'all',
-                     desc: 'OPTIONAL: Filter unallocated claims. Some filters only apply to AGFS or LGFS schemas.'
+            optional  :status,
+                      type: String,
+                      default: 'current',
+                      values: %w[current allocated unallocated archived],
+                      desc: 'REQUIRED: Only returns claims in the specified status'
+            optional  :scheme,
+                      type: String,
+                      default: 'agfs',
+                      values: %w[agfs lgfs],
+                      desc: 'OPTIONAL: This will be used to filter the list of allocated/unallocated claims'
+            optional  :filter,
+                      type: String,
+                      default: 'all',
+                      values: %w[all redetermination awaiting_written_reasons fixed_fee cracked trial guilty_plea graduated_fees interim_fees warrants interim_disbursements risk_based_bills],
+                      desc: 'OPTIONAL: Filter unallocated claims. Some filters only apply to AGFS or LGFS schemas.'
             use :searching
             use :sorting
             use :pagination
@@ -78,16 +77,7 @@ module API
             end
 
             def claims_scope
-              case params.status
-              when 'current'
-                current_claims
-              when 'allocated'
-                allocated_claims
-              when 'unallocated'
-                unallocated_claims
-              when 'archived'
-                archived_claims
-              end
+              send("#{params.status}_claims")
             end
 
             def claims
